fixed the header, and added the structure for main part
Added the secont part of the header

Put all the templating for the main page so it can be easaly seperated

added the caroussel, just a placeholder for the momet

added the 3 blog sections. Just a place holder for the moment.

Need the sidebar and footer
also need some css foo on the header